+++ CHANGELOG.md
@@ -1,11 +1,116 @@
-## v0.0.3 (2021-07-26)
+## v0.0.6 (2021-07-28)
 
 ### Fix
 
-- **main.py**: test commitizen
+- nothing
 
-## v0.0.2 (2021-07-26)
+## v0.0.5 (2021-07-28)
 
 ### Fix
 
-- **main.py**: commitize2 not commitizen
+- nothing
+
+## v0.0.5 (2021-07-28)
+
+### Fix
+
+- nothing
+- nothing
+- nothing
+
+## v0.0.4 (2021-07-28)
+
+### Fix
+
+- nothing
+
+## v0.0.5 (2021-07-28)
+
+### Fix
+
+- nothing
+- nothing
+- nothing
+
+## v0.0.4 (2021-07-28)
+
+### Fix
+
+- nothing
+- print error
+- char map
+- nothing
+
+### Perf
+
+- nothing
+
+## v0.0.3 (2021-07-28)
+
+### Fix
+
+- nothing
+
+## v0.0.5 (2021-07-28)
+
+### Fix
+
+- nothing
+- nothing
+- nothing
+
+## v0.0.4 (2021-07-28)
+
+### Fix
+
+- nothing
+- nothing
+- nothing
+
+## v0.0.4 (2021-07-28)
+
+### Fix
+
+- nothing
+- nothing
+
+## v0.0.4 (2021-07-28)
+
+### Fix
+
+- nothing
+
+## v0.0.4 (2021-07-28)
+
+### Fix
+
+- nothing
+- print error
+- char map
+- nothing
+
+### Perf
+
+- nothing
+
+## v0.0.3 (2021-07-27)
+
+### Perf
+
+- nothing
+
+## v0.0.3 (2021-07-27)
+
+### Fix
+
+- **add_commit.py**: fix problemm with cz bump command
+- **add_commit.py**: fixing a problem with parsing the output of one of the pipenv run cz bump -ch
+
+## v0.0.2 (2021-07-27)
+
+### Fix
+
+- **add_commit.py**: fix problemm with cz bump command
+- **add_commit.py**: fixing a problem with parsing the output of one of the pipenv run cz bump -ch
+
+## v0.0.0.alpha (2021-07-26)